css/created
Created some basic css. It currently has a mix of grid and flex. It is not fully responsive just yet becaause there are a lot of defined pixels. I tried to get it to match the mock site as much as possible, but I think this will do for a rough draft