rooms, carousel and reservation
I fixed the CSS and added lorem text to the rooms information. I also updated the code for the carousel so it should run smoother especially with the new pics added. I also moved the agree to terms and conditions check box to the end of the form on the reservation page