Functionality
Added Terms and conditions to the reseration page and made a few astetic changes. It shows a change to teh server.js but that's incorrect. I was going to add the popper code for the terms and conditions but moved it to the easter egg js file instead to keep from crashing our site